third_party/blink/web_tests/platform/generic/http/tests/pointer-lock/pointerlockelement-different-origin-expected.txt

Test iframe from different origin can not access pointerLockElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

     Lock targetDiv1.
PASS onpointerlockchange received after: Lock targetDiv1.
PASS document.pointerLockElement is targetDiv1
PASS message is "inner-iframe.html document.pointerLockElement = null"
PASS successfullyParsed is true

TEST COMPLETE
